+++ sshd-common/src/main/java/org/apache/sshd/common/cipher/BuiltinCiphers.java
@@ -31,6 +31,7 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeMap;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -39,6 +40,7 @@ import org.apache.sshd.common.NamedResource;
 import org.apache.sshd.common.config.NamedFactoriesListParseResult;
 import org.apache.sshd.common.util.GenericUtils;
 import org.apache.sshd.common.util.ValidateUtils;
+import org.apache.sshd.common.util.security.SecurityUtils;
 
 /**
  * Provides easy access to the currently implemented ciphers
@@ -50,6 +52,11 @@ public enum BuiltinCiphers implements CipherFactory {
         @Override
         public Cipher create() {
             return new CipherNone();
+        }
+
+        @Override
+        public boolean isSupported() {
+            return !SecurityUtils.isFipsMode();
         }
     },
     aes128cbc(Constants.AES128_CBC, 16, 0, 16, "AES", 128, "AES/CBC/NoPadding", 16) {
@@ -149,6 +156,11 @@ public enum BuiltinCiphers implements CipherFactory {
         public Cipher create() {
             return new ChaCha20Cipher();
         }
+
+        @Override
+        public boolean isSupported() {
+            return !SecurityUtils.isFipsMode();
+        }
     },
     /**
      * @deprecated
@@ -175,7 +187,7 @@ public enum BuiltinCiphers implements CipherFactory {
     private final int blkSize;
     private final String algorithm;
     private final String transformation;
-    private final boolean supported;
+    private final AtomicReference<Boolean> supported = new AtomicReference<>();
 
     BuiltinCiphers(
                    String factoryName, int ivsize, int authSize, int kdfSize,
@@ -188,13 +200,6 @@ public enum BuiltinCiphers implements CipherFactory {
         this.algorithm = algorithm;
         this.transformation = transformation;
         this.blkSize = blkSize;
-        /*
-         * This can be done once since in order to change the support the JVM needs to be stopped, some
-         * unlimited-strength files need be installed and then the JVM re-started. Therefore, the answer is not going to
-         * change while the JVM is running
-         */
-        this.supported = Constants.NONE.equals(factoryName) || Constants.CC20P1305_OPENSSH.equals(factoryName)
-                || Cipher.checkSupported(this.transformation, this.keysize);
     }
 
     @Override
@@ -213,7 +218,14 @@ public enum BuiltinCiphers implements CipherFactory {
      */
     @Override
     public boolean isSupported() {
-        return supported;
+        Boolean value = supported.get();
+        if (value == null) {
+            value = Cipher.checkSupported(this.transformation, this.keysize);
+            if (!supported.compareAndSet(null, value)) {
+                value = supported.get();
+            }
+        }
+        return value.booleanValue();
     }
 
     @Override